model: tensorflow: updated to support archive storage
